fixed bugs

+++ pepecontroller/src/commands/vote.h
@@ -1,6 +1,6 @@
 #pragma once
 #include "command.h"
-#include "phase.h"
+#include "votePhase.h"
 #include <unordered_set>
 #include <cstdint>
 
@@ -11,11 +11,11 @@ class Vote : public Command
 public:
 	//void resetCache();
 	/*Vote(uint32_t num);*/
-	Vote(VotePhase* ph, uint8_t num);
+	Vote(VotePhase* ph, uint32_t num);
 	void Execute(Context* ctx) override;
 private:
 	VotePhase* phase;
-	uint8_t num;
+	uint32_t num;
 	//std::unordered_set<size_t> voteCache;
 };
