+++ pepecontroller/src/commands/voteConstructor.cpp
@@ -1,10 +1,12 @@
 #include "voteConstructor.h"
+#include "votePhase.h"
 #include "vote.h"
 #include <sstream>
 #include <limits>
 
 namespace commands
 {
+
 	VoteConstructor::VoteConstructor(VotePhase* ph) : phase(ph) {};
 
 std::unique_ptr<Command> VoteConstructor::Construct(const peka2tv::Peka2tvSIOClient::ChatMessage& msg)
@@ -12,7 +14,7 @@ std::unique_ptr<Command> VoteConstructor::Construct(const peka2tv::Peka2tvSIOCli
 	std::stringstream ss(msg.text);
 	ss.ignore(std::numeric_limits<std::streamsize>::max(), ' '); // ignore text before space
 
-	uint8_t num;
+	uint32_t num;
 	ss >> num;
 
 	return std::make_unique<Vote>(phase, num);
